Changed main package from corpusloader to atap_corpus_loader
