+++ atap_corpus_loader/controller/FileLoaderService.py
@@ -4,8 +4,8 @@ from typing import Optional
 from atap_corpus.corpus.corpus import DataFrameCorpus
 from pandas import DataFrame, merge, concat
 
-from corpusloader.controller.data_objects import FileReference, CorpusHeader, ZipFileReference
-from corpusloader.controller.file_loader_strategy import FileLoaderStrategy, FileLoaderFactory, FileLoadError
+from atap_corpus_loader.controller.data_objects import FileReference, CorpusHeader, ZipFileReference
+from atap_corpus_loader.controller.file_loader_strategy import FileLoaderStrategy, FileLoaderFactory, FileLoadError
 
 
 class FileLoaderService:
