+++ atap_corpus_loader/view/gui/AbstractWidget.py
