+++ atap_corpus_loader/view/gui/OniLoaderWidget.py
@@ -1,7 +1,7 @@
 from panel import Row
 
-from corpusloader.controller import Controller
-from corpusloader.view.gui import AbstractWidget
+from atap_corpus_loader.controller import Controller
+from atap_corpus_loader.view.gui import AbstractWidget
 
 
 class OniLoaderWidget(AbstractWidget):
